enemy spawn frequency based on intensity
Enemies now spawn more often when the intensity is higher.
Also, the mechanism has been modified to only use the "chance spawn enemy" function but used a counter so that a regular rhythm of enemies is always present but additional enemies are added randomly with increased frequency at higher intensities (these random enemies do not occur in the first 10 or so seconds).